refactor: 로또 번호 생성 전략패턴 적용
