+++ src/main/java/lotto/LottoMainApplication.java
@@ -1,11 +1,11 @@
 package lotto;
 
 import lotto.domain.LotteryShop;
-import lotto.domain.LottoNumberGenerator;
 import lotto.domain.LottoTickets;
 import lotto.domain.Money;
 import lotto.domain.Statistics;
 import lotto.domain.WinNumbers;
+import lotto.domain.strategy.RandomNumbers;
 import lotto.ui.InputView;
 import lotto.ui.ResultView;
 
@@ -14,7 +14,7 @@ public class LottoMainApplication {
   public static void main(String[] args) {
     Money amount = Money.of(InputView.requestPurchaseAmount());
     LotteryShop lotteryShop = new LotteryShop();
-    LottoTickets lottoTickets = lotteryShop.sell(amount, new LottoNumberGenerator());
+    LottoTickets lottoTickets = lotteryShop.sell(amount, new RandomNumbers());
     ResultView.printLottoCount(lottoTickets.size());
     ResultView.printLottoTickets(lottoTickets);
