+++ src/main/java/lotto/domain/strategy/GenerateNumbersStrategy.java
@@ -1,0 +1,8 @@
+package lotto.domain.strategy;
+
+import java.util.List;
+
+public interface GenerateNumbersStrategy {
+
+  List<Integer> generate();
+}
