+++ src/main/java/lotto/domain/strategy/RandomNumbers.java
@@ -1,4 +1,4 @@
-package lotto.domain;
+package lotto.domain.strategy;
 
 import static java.util.stream.IntStream.rangeClosed;
 
@@ -6,7 +6,7 @@ import java.util.Collections;
 import java.util.List;
 import java.util.stream.Collectors;
 
-public class LottoNumberGenerator {
+public class RandomNumbers implements GenerateNumbersStrategy {
 
   private static final int FROM_INDEX = 0;
   private static final int TO_INDEX = 6;
