+++ src/test/java/lotto/domain/strategy/RandomNumbersTest.java
@@ -1,4 +1,4 @@
-package lotto.domain;
+package lotto.domain.strategy;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
@@ -6,13 +6,13 @@ import java.util.List;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.RepeatedTest;
 
-public class LottoNumberGeneratorTest {
+public class RandomNumbersTest {
 
   @DisplayName("중복없는 로또 번호 6개가 반환된다")
   @RepeatedTest(100)
   void generateLottoNumber() {
-    LottoNumberGenerator lottoNumberGenerator = new LottoNumberGenerator();
-    List<Integer> generate = lottoNumberGenerator.generate();
+    GenerateNumbersStrategy generateNumbersStrategy = new RandomNumbers();
+    List<Integer> generate = generateNumbersStrategy.generate();
     assertThat(generate).hasSize(6);
     assertThat(generate).doesNotHaveDuplicates();
     assertThat(generate).hasSizeBetween(1, 45);
